[vfs] Keep a stack of offsets in url.
Suppose a following situation: user goes into second page of some directory A,
choses a subdirectory B, goes into B and then clicks an 'Up' link. The question
is which page of directory A will be shown. Whithout any support it of course
will be the first page, but assuming user were brousing the second page such
behaviour is annoying to user.

To fix this we keep a stack of offsets in url. The latter means there is no
longer only one 'o' GET parameter but a whole bunch of them for each directory
in user's digging history.