[search] Move usertypes to common.py. More query extensions
Add net, port, order query extensions. For order extension it is possible to
have up to 3 keys separated by comma: state (online/offline), host (key is
network + netshare_id so it is not quite the same as hostname) and size (file
size.
The key got from type query extension now is passed directly to database with
only one exception 'type:dir' which is handled as earlier.