[network.py] Open port scanner in python
Due to nmap ridiculous license we have to decline using it as a scanner for
our lookuper and pinger.

scan_hosts() accepts a list of parts (host, port) and return a list of online
hosts with corresponding ports opened. It is select()-based for portability
reasons.

also a small example is provided as __main__