[libuguu] Do not use per-node stamps in the tree data structure anymore
d->stamp have been used to determine if we should scan a node or leave it.
This can be done with O(1) memory instead of O(n) using the following simple
rule: the node is to be scanned if we've come to it from it's parent or
sibling, if we've come from a child the node shoul be left.