[DDL] Do not rely on filenames table for search anymore
Store file names in the files table.
Write path tsvector for a file in the time a line for the file is inserted
into the files table, so we get rid of a very slow UPDATE command in the
spider.

Global rss feeds are disable since they are depended on the filenames table.
Unless noone objects global rss feeds will be removed in future